requirements.txt
datasets
tqdm
pandas
numpy
zstandard
python-chess
